requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pydantic
google-genai
python-docx
reportlab
pypdf
pdfplumber
aiofiles
jinja2
pydantic[email]
